refactor(opencode): extract the launch gate result from the runtime adapter
Problem: the shape of a refused OpenCode launch is decided in the middle of the
runtime adapter. `blockedLaunchResult` and the four helpers that pick the text
it shows a user - the generic-reason list, the secret redaction, the timing-line
filter, the first-displayable-diagnostic walk - sit near the bottom of a
1564-line file, several hundred lines below the twelve call sites that build a
block from them, and below the retry policy that has to reason about whether a
block is safe to relaunch. `OpenCodeLaunchGateResult.ts` already holds the other
half of that decision: which reasons carry a pre-launch gate marker, and which
of those a caller may retry on its own. The two halves cannot be read together,
and the module that owns the retry rule cannot construct the result the rule
applies to.

Fix: move the result builder and its message helpers into the module that
already owns the gate rules. Nothing changes behaviourally - the adapter now
imports `blockedLaunchResult` and the three helpers it still calls directly
instead of declaring them - and the launch-gate unit test grows the cases that
were previously reachable only through a full adapter launch: the readiness
reason that mines the diagnostics for a displayable message, the non-readiness
reason reported verbatim, the opt-in gate marker, secret redaction, and the
generic/timing filters.

Why this shape: the gate marker and the result it is attached to are one
decision, so they belong in one file. Keeping `blockedLaunchResult` in the
adapter would mean the retry policy imports a rule from a module that imports
the result builder back, and it is the reason the marker rule was previously
duplicated as an inline object literal. `isGenericOpenCodeFailureMessage` stays
module-private: it exists to serve the walk above it, and no caller outside this
file has ever asked the question it answers.
